Eliminate numba warnings and move TileCalculator class to another file
